OfficeXlsFile2 linux build - complete
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@64139 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficeXlsFile2/source/XlsFormat/Logic/Biff_records/MergeCells.cpp
@@ -48,7 +48,7 @@ int MergeCells::serialize(std::wostream & stream)
 	CP_XML_WRITER(stream)    
     {
 		
-		for (long i = 0 ; i < rgref.size(); i++)
+        for (int i = 0 ; i < rgref.size(); i++)
 		{	
 			Ref8* ref = dynamic_cast<Ref8*>(rgref[i].get());
 			CP_XML_NODE(L"mergeCell")
