+++ ASCOfficeXlsFile2/source/XlsFormat/Logic/Biff_structures/CellXF.cpp
@@ -114,7 +114,7 @@ void CellXF::load(CFRecord& record)
 
 void CellXF::RegisterFillBorder()
 {
-	for (long i = 0; i < ext_props.size(); i++ )
+    for (int i = 0; i < ext_props.size(); i++ )
 	{
 		ExtProp* ext_prop = dynamic_cast<ExtProp*>(ext_props[i].get());
