+++ ASCOfficeXlsFile2/source/XlsFormat/Logic/Biff_structures/SortCond12.h
@@ -28,7 +28,7 @@ public:
 	CondDataValue condDataValue;
 	CFFlag cfflag;
 
-	long cchSt;
+    int cchSt;
 	XLUnicodeStringNoCch stSslist;
 };
