+++ ASCOfficeXlsFile2/source/XlsFormat/Logic/Biff_unions/STYLES.cpp
@@ -87,7 +87,7 @@ int STYLES::serialize(std::wostream & stream)
 									}
 								}
 
-								for (long i = 0; i < styleExt->xfProps.xfPropArray.size(); i++)
+                                for (int i = 0; i < styleExt->xfProps.xfPropArray.size(); i++)
 								{
 								}
 							}
